+++ src/main/java/fr/xephi/authme/command/CommandHandler.java
@@ -3,6 +3,9 @@ package fr.xephi.authme.command;
 import java.util.ArrayList;
 import java.util.List;
 
+import fr.xephi.authme.AuthMe;
+import fr.xephi.authme.command.help.HelpProvider;
+import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
 
 import fr.xephi.authme.util.StringUtils;
@@ -12,6 +15,12 @@ import fr.xephi.authme.util.StringUtils;
  * or to display help messages for unknown invocations.
  */
 public class CommandHandler {
+
+    /**
+     * The threshold for suggesting a similar command. If the difference is below this value, we will
+     * ask the player whether he meant the similar command.
+     */
+    private static final double SUGGEST_COMMAND_THRESHOLD = 0.75;
 
     private final CommandService commandService;
 
@@ -40,12 +49,30 @@ public class CommandHandler {
         parts.add(0, bukkitCommandLabel);
 
         FoundCommandResult result = commandService.mapPartsToCommand(sender, parts);
-        if (FoundResultStatus.SUCCESS.equals(result.getResultStatus())) {
-            executeCommand(sender, result);
-        } else {
-            commandService.outputMappingError(sender, result);
-        }
+        handleCommandResult(sender, result);
         return !FoundResultStatus.MISSING_BASE_COMMAND.equals(result.getResultStatus());
+    }
+
+    private void handleCommandResult(CommandSender sender, FoundCommandResult result) {
+        switch (result.getResultStatus()) {
+            case SUCCESS:
+                executeCommand(sender, result);
+                break;
+            case MISSING_BASE_COMMAND:
+                sender.sendMessage(ChatColor.DARK_RED + "Failed to parse " + AuthMe.getPluginName() + " command!");
+                break;
+            case INCORRECT_ARGUMENTS:
+                sendImproperArgumentsMessage(sender, result);
+                break;
+            case UNKNOWN_LABEL:
+                sendUnknownCommandMessage(sender, result);
+                break;
+            case NO_PERMISSION:
+                sendPermissionDeniedError(sender);
+                break;
+            default:
+                throw new IllegalStateException("Unknown result status '" + result.getResultStatus() + "'");
+        }
     }
 
     /**
@@ -76,6 +103,46 @@ public class CommandHandler {
         return cleanArguments;
     }
 
+    /**
+     * Show an "unknown command" message to the user and suggest an existing command if its similarity is within
+     * the defined threshold.
+     *
+     * @param sender The command sender
+     * @param result The command that was found during the mapping process
+     */
+    private static void sendUnknownCommandMessage(CommandSender sender, FoundCommandResult result) {
+        sender.sendMessage(ChatColor.DARK_RED + "Unknown command!");
 
+        // Show a command suggestion if available and the difference isn't too big
+        if (result.getDifference() <= SUGGEST_COMMAND_THRESHOLD && result.getCommandDescription() != null) {
+            sender.sendMessage(ChatColor.YELLOW + "Did you mean " + ChatColor.GOLD
+                + CommandUtils.constructCommandPath(result.getCommandDescription()) + ChatColor.YELLOW + "?");
+        }
+
+        sender.sendMessage(ChatColor.YELLOW + "Use the command " + ChatColor.GOLD + "/" + result.getLabels().get(0)
+            + " help" + ChatColor.YELLOW + " to view help.");
+    }
+
+    private void sendImproperArgumentsMessage(CommandSender sender, FoundCommandResult result) {
+        CommandDescription command = result.getCommandDescription();
+        if (!commandService.getPermissionsManager().hasPermission(sender, command)) {
+            sendPermissionDeniedError(sender);
+            return;
+        }
+
+        // Show the command argument help
+        sender.sendMessage(ChatColor.DARK_RED + "Incorrect command arguments!");
+        commandService.outputHelp(sender, result, HelpProvider.SHOW_ARGUMENTS);
+
+        List<String> labels = result.getLabels();
+        String childLabel = labels.size() >= 2 ? labels.get(1) : "";
+        sender.sendMessage(ChatColor.GOLD + "Detailed help: " + ChatColor.WHITE
+            + "/" + labels.get(0) + " help " + childLabel);
+    }
+
+    // TODO ljacqu 20151212: Remove me once I am a MessageKey
+    private static void sendPermissionDeniedError(CommandSender sender) {
+        sender.sendMessage(ChatColor.DARK_RED + "You don't have permission to use this command!");
+    }
 
 }
